backend/pyproject.toml
[project]
name = "cocoon-backend"
version = "0.1.0"
description = "Microsoft SoW Automation - Cocoon Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "python-arango>=8.1.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
